# protocol/validation: cache per-entry validation results during ValidateTx