A new benchmark (included in this PR) led us to discover that during `ValidateTx` of certain kinds of very large transactions, some entries were being validated multiple times - in some cases, thousands of times each. This short-circuits that by caching the result of entry validation for the duration of `ValidateTx`.

```
$ benchstat before.txt after.txt 
name                                                   old time/op  new time/op  delta
BuildTx/1-asset--1-issuance--build-8                    591µs ± 1%   606µs ± 3%     ~     (p=0.056 n=5+5)
BuildTx/1-asset--1-issuance--build-sign-8               852µs ± 1%   838µs ± 0%   -1.56%  (p=0.008 n=5+5)
BuildTx/1-asset--1-issuance--build-sign-finalize-8      339ms ± 7%   333ms ±12%     ~     (p=0.841 n=5+5)
BuildTx/1-asset--10-issuance--build-8                   790µs ± 2%   776µs ± 0%   -1.68%  (p=0.008 n=5+5)
BuildTx/1-asset--10-issuance--build-sign-8             2.88ms ± 1%  3.23ms ± 8%  +12.28%  (p=0.008 n=5+5)
BuildTx/1-asset--10-issuance--build-sign-finalize-8     325ms ± 0%   333ms ± 2%     ~     (p=0.190 n=4+5)
BuildTx/1-asset--100-issuance--build-8                 2.78ms ± 1%  2.94ms ±10%     ~     (p=0.548 n=5+5)
BuildTx/1-asset--100-issuance--build-sign-8            21.2ms ± 1%  24.4ms ± 2%  +15.11%  (p=0.008 n=5+5)
BuildTx/1-asset--100-issuance--build-sign-finalize-8    361ms ± 1%   363ms ± 1%     ~     (p=0.222 n=5+5)
BuildTx/10-asset--1-issuance--build-8                  2.48ms ± 2%  2.49ms ± 1%     ~     (p=0.421 n=5+5)
BuildTx/10-asset--1-issuance--build-sign-8             4.46ms ± 1%  4.50ms ± 1%     ~     (p=0.056 n=5+5)
BuildTx/10-asset--1-issuance--build-sign-finalize-8     345ms ± 0%   333ms ± 2%   -3.60%  (p=0.016 n=4+5)
BuildTx/10-asset--10-issuance--build-8                 4.41ms ± 1%  4.44ms ± 1%     ~     (p=0.310 n=5+5)
BuildTx/10-asset--10-issuance--build-sign-8            23.2ms ± 2%  23.6ms ± 1%   +1.68%  (p=0.008 n=5+5)
BuildTx/10-asset--10-issuance--build-sign-finalize-8    531ms ± 5%   373ms ± 2%  -29.83%  (p=0.008 n=5+5)
BuildTx/10-asset--100-issuance--build-8                23.2ms ± 1%  23.6ms ± 2%     ~     (p=0.222 n=5+5)
BuildTx/10-asset--100-issuance--build-sign-8            211ms ± 2%   211ms ± 1%     ~     (p=0.690 n=5+5)
BuildTx/10-asset--100-issuance--build-sign-finalize-8   2.32s ± 2%   0.75s ± 2%  -67.46%  (p=0.008 n=5+5)
BuildTx/100-asset--1-issuance--build-8                 19.6ms ± 1%  19.7ms ± 1%     ~     (p=0.548 n=5+5)
BuildTx/100-asset--1-issuance--build-sign-8            38.1ms ± 1%  37.9ms ± 2%     ~     (p=0.310 n=5+5)
BuildTx/100-asset--1-issuance--build-sign-finalize-8    2.12s ± 3%   0.42s ± 1%  -80.35%  (p=0.008 n=5+5)
BuildTx/100-asset--10-issuance--build-8                38.8ms ± 1%  39.3ms ± 3%     ~     (p=0.841 n=5+5)
BuildTx/100-asset--10-issuance--build-sign-8            224ms ± 2%   227ms ± 1%     ~     (p=0.421 n=5+5)
BuildTx/100-asset--10-issuance--build-sign-finalize-8   18.5s ± 2%    0.9s ± 2%  -95.27%  (p=0.008 n=5+5)
```