@bobg 2017-05-18
```
$ go test -run qqq -bench BenchmarkBuildTx/100-asset--10-issuance--build-sign-finalize ./core
```

Before: 20,944,675,820 ns/op
After: 876,369,721 ns/op