@kr 2017-05-19
Do you mind pasting the output of https://godoc.org/golang.org/x/perf/cmd/benchstat in the commit message?