@jbowens 2017-05-19
To get useful output out of benchstat (with statistically significant deltas), you might need to run the benchmarks a few time on both old and new code. There could be an easier way, but I just `>>` pipe the output of go test into the same file five times for each. It'd be nice if `go test` had an option for this.

https://godoc.org/rsc.io/benchstat#hdr-Example