@jbowens 2017-05-19
>  It'd be nice if go test had an option for this.

Ah, maybe you can use the `-count` flag? https://github.com/golang/go/blob/master/src/cmd/go/internal/test/test.go#L157-L160