@kr 2017-05-19
Noice. -95.27% is music to my eyes.